Addition of the beginnings of the Game project
This commit adds quite a bit to the git folder. Included are multiple development projects used to ensure the functionality of directX. The main commit resides around the implementation of the Molinator (very initial name) game. WASD, X, and SPACE are used to control the environment. To use the game, you will most likely need to install the directX9 SDK, the FBX SDK, and the Windows SDK.